fix: gnome 44 install fails bc no schemas folder
When installing gnome 44 expects a schemas folder to be present, even
if that folder is empty. If the folder is missing then an error is logged
and the installation terminates uncessfully. I have reported this at [1].

This commit adds an empty schema's folder so that installation is successful
until a fix is released or the docs make folder an explicit requirement.

[1] https://bugs.launchpad.net/ubuntu/+source/gnome-shell/+bug/2013073